+++ app/code/Magento/InventoryConfiguration/Controller/Adminhtml/Report/ExportLowstockCsv.php
@@ -4,12 +4,14 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\InventoryConfiguration\Controller\Adminhtml\Report\Product;
+
+namespace Magento\InventoryConfiguration\Controller\Adminhtml\Report;
 
 use Magento\Framework\App\ResponseInterface;
 use Magento\Framework\App\Filesystem\DirectoryList;
+use Magento\Reports\Controller\Adminhtml\Report\Product as ProductReportController;
 
-class ExportLowstockExcel extends \Magento\Reports\Controller\Adminhtml\Report\Product
+class ExportLowstockCsv extends ProductReportController
 {
     /**
      * Authorization level of a basic admin session
@@ -19,21 +21,23 @@ class ExportLowstockExcel extends \Magento\Reports\Controller\Adminhtml\Report\P
     const ADMIN_RESOURCE = 'Magento_Reports::report_products';
 
     /**
-     * Export low stock products report to XML format
+     * Export low stock products report to CSV format
      *
      * @return ResponseInterface
      */
     public function execute()
     {
         $this->_view->loadLayout(false);
-        $fileName = 'products_lowstock.xml';
+        $fileName = 'products_lowstock.csv';
+
         $exportBlock = $this->_view->getLayout()->getChildBlock(
-            'adminhtml.block.report.product.inventory.grid',
+            'adminhtml.block.report.product.inventory.lowstock.grid',
             'grid.export'
         );
+
         return $this->_fileFactory->create(
             $fileName,
-            $exportBlock->getExcelFile(),
+            $exportBlock->getCsvFile(),
             DirectoryList::VAR_DIR
         );
     }
