+++ app/code/Magento/InventoryConfiguration/Controller/Adminhtml/Report/ExportLowstockExcel.php
@@ -4,13 +4,14 @@
  * Copyright © Magento, Inc. All rights reserved.
  * See COPYING.txt for license details.
  */
-namespace Magento\InventoryConfiguration\Controller\Adminhtml\Report\Product;
+
+namespace Magento\InventoryConfiguration\Controller\Adminhtml\Report;
 
 use Magento\Framework\App\ResponseInterface;
 use Magento\Framework\App\Filesystem\DirectoryList;
-use Magento\Reports\Controller\Adminhtml\Report\Product as ProductReportController;
+use Magento\Reports\Controller\Adminhtml\Report\Product  as ProductReportController;
 
-class ExportLowstockCsv extends ProductReportController
+class ExportLowstockExcel extends ProductReportController
 {
     /**
      * Authorization level of a basic admin session
@@ -20,21 +21,23 @@ class ExportLowstockCsv extends ProductReportController
     const ADMIN_RESOURCE = 'Magento_Reports::report_products';
 
     /**
-     * Export low stock products report to CSV format
+     * Export low stock products report to XML format
      *
      * @return ResponseInterface
      */
     public function execute()
     {
         $this->_view->loadLayout(false);
-        $fileName = 'products_lowstock.csv';
+        $fileName = 'products_lowstock.xml';
+
         $exportBlock = $this->_view->getLayout()->getChildBlock(
-            'adminhtml.block.report.product.inventory.grid',
+            'adminhtml.block.report.product.inventory.lowstock.grid',
             'grid.export'
         );
+
         return $this->_fileFactory->create(
             $fileName,
-            $exportBlock->getCsvFile(),
+            $exportBlock->getExcelFile(),
             DirectoryList::VAR_DIR
         );
     }
